requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
PyJWT==2.11.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
gunicorn==22.0.0